Updated event controller to ensure the user is authorized.
